refactor(agent): split Codex adopted goal turn
